+++ build.rs
@@ -1,6 +1,22 @@
-// We need this build script to handle Thumb mode for ARM on a stable
-// release channel, where `target_feature = "thumb_mode"` isn't available.
-fn main() {
+use rustflags::Flag;
+
+/// Ensure that the target arch is supported.
+///
+/// Doing this here instead of emitting `compile_error!` in the lib itself leads to better error
+/// messages.
+fn check_supported_archs() {
+    let arch = std::env::var("CARGO_CFG_TARGET_ARCH").unwrap();
+
+    if !["x86_64", "x86", "aarch64", "arm"].contains(&arch.as_str()) {
+        println!("cargo::error=closure-ffi does not support the '{arch}' target architecture.");
+    }
+}
+
+/// Set a `thumb_mode` cfg on arm targets using thumb encoding by default.
+///
+/// We need this build script to handle Thumb mode for ARM on a stable
+/// release channel, where `target_feature = "thumb_mode"` isn't available.
+fn set_thumb_mode_cfg() {
     println!("cargo::rustc-check-cfg=cfg(thumb_mode)");
 
     let arch = std::env::var("CARGO_CFG_TARGET_ARCH").unwrap();
@@ -18,7 +34,7 @@ fn main() {
             // so fall back on assuming the feature simply isn't set
             if target.to_lowercase().ends_with(".json") {
                 println!(
-                    r"cargo::warning=Custom ARM target file detected. If using the Thumb ISA by 
+                    "cargo::warning=Custom ARM target file detected. If using the Thumb ISA by \
                     default, use nightly and add 'thumb-mode' to the feature list."
                 );
                 false
@@ -30,4 +46,27 @@ fn main() {
     {
         println!("cargo::rustc-cfg=thumb_mode")
     }
+}
+
+// Make sure that the unstable feature is enabled when `-C instrument-coverage` is passed to rustc.
+// We need it to add #[coverage(off)] on compiler generated thunks
+fn check_coverage_supported() {
+    if std::env::var("CARGO_FEATURE_COVERAGE").is_ok() {
+        return;
+    }
+
+    if rustflags::from_env()
+        .any(|f| matches!(f, Flag::Codegen { opt, value: _ } if opt == "instrument-coverage"))
+    {
+        println!(
+            "cargo::error=closure-ffi requires a nightly compiler and the 'coverage' crate feature \
+            to be enabled for '-C instrument-coverage' support."
+        );
+    }
+}
+
+fn main() {
+    check_supported_archs();
+    check_coverage_supported();
+    set_thumb_mode_cfg();
 }